pyproject.toml
[tool.poetry]
name = "agent-factory"
version = "0.1.0"
description = "A scalable framework for creating specialized AI agents with dynamic tool assignment"
authors = ["Agent Factory Contributors"]
readme = "README.md"
license = "MIT"
keywords = ["langchain", "agents", "ai", "llm", "automation"]
package-mode = false  # Poetry 2.x: This is a framework/application, not a library package

[tool.poetry.dependencies]
python = ">=3.11.0,<3.13"

# Core LangChain dependencies
langchain = "^0.2.1"
langchain-openai = "^0.1.8"
langchain-community = "^0.2.1"
langchain-anthropic = "^0.1.15"
langchain-google-genai = "^1.0.5"
langchainhub = "^0.1.18"

# LLM & Embeddings
tiktoken = "^0.7.0"

# Web search and research tools
wikipedia = "^1.4.0"
tavily-python = "^0.3.3"
duckduckgo-search = "^4.1.0"

# File operations and Git
gitpython = "^3.1.40"

# Utilities
python-dotenv = "^1.0.1"
rich = "^13.7.0"
tenacity = "^8.2.3"

# CLI tools
typer = "^0.12.0"
prompt-toolkit = "^3.0.43"

# Code generation
jinja2 = "^3.1.2"
markdown = "^3.5.0"

# File format support
PyPDF2 = "^3.0.1"

# Data structures
pydantic = "^2.5.0"
networkx = ">=3.2,<3.6"

# GitHub integration
pygithub = "^2.1.1"
pyyaml = "^6.0"
litellm = "1.30.0"
fastapi = "^0.124.0"
uvicorn = {extras = ["standard"], version = "^0.38.0"}
python-multipart = "^0.0.20"

# Telegram bot integration
python-telegram-bot = "^22.5"
supabase = "^2.25.0"
edge-tts = "^7.2.3"
moviepy = "^2.2.1"
pillow = "11.0.0"
filelock = "^3.13.0"  # Cross-platform process lock to prevent duplicate bot instances
aiohttp = "^3.9.0"  # HTTP server for health check endpoint
psycopg2-binary = "^2.9.11"
psycopg = {extras = ["binary"], version = "^3.3.2"}
psycopg-pool = "^3.3.0"

[tool.poetry.group.dev.dependencies]
# Development tools
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.12.0"
isort = "^5.13.0"
ruff = "^0.1.8"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [
    ".cache",
    ".venv",
    "**/__pycache__",
    "Agent-Factory",     # Removed duplicate directory
    "crews",             # Generated YAML specs
    "scripts",           # Shell scripts
    "tests",             # Test files
    ".pytest_cache",
    "dist",
    "build",
    ".git",
    ".githooks",
    "drafts"
]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']
line-length = 100

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
# Ignore removed duplicate directory and cache
norecursedirs = ["Agent-Factory", ".git", ".venv", "__pycache__", ".pytest_cache", "drafts"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.poetry.scripts]
agentcli = "agentcli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
